possible fix for migration error

+++ database/migrations/2020_12_31_200032_create_user_dividends.php
@@ -15,8 +15,8 @@ class CreateUserDividends extends Migration
     {
         Schema::create('user_dividends', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id', false, true);
-            $table->integer('dividend_id', false, true);
+            $table->bigInteger('user_id', false, true);
+            $table->bigInteger('dividend_id', false, true);
             $table->float('quantity')->default(0.00);
             $table->float('portfolio_value')->default(0.00);
             $table->timestamps();
